requirements.txt
Django==5.1.3
openai==1.57.1
Pillow==11.0.0
Requests==2.32.3
